requirements.txt
aiogram==3.13.1
telethon==1.36.0
python-dotenv==1.0.1
requests==2.31.0
rich==13.9.2
httpx==0.27.0
